+++ bindgen/Cargo.toml
@@ -15,6 +15,7 @@ path = "src/main.rs"
 anyhow = "1"
 askama = { version = "0.11", default-features = false, features = ["config"] }
 camino = "1.0.8"
+cargo_metadata = "0.15"
 clap = { version = "3.1", features = ["cargo", "std", "derive"] }
 extend = "1.1"
 fs-err = "2.7.0"
@@ -24,4 +25,7 @@ serde = "1"
 serde_json = "1.0.0"
 textwrap = "0.16"
 toml = "0.5"
-uniffi_bindgen = { path = "../3rd-party/uniffi-rs/uniffi_bindgen" }
+
+uniffi_bindgen.workspace = true
+uniffi_meta.workspace = true
+uniffi_udl.workspace = true
