+++ bindgen/templates/macros.cs
@@ -30,6 +30,17 @@
         {{- prefix }}, {% call lower_arg_list(func) -%}{% if func.arguments().len() > 0 %},{% endif %} ref _status)
 )
 {%- endmacro -%}
+{%- macro to_ffi_method_call(func) %}
+    {%- match func.throws_type() %}
+    {%- when Some with (e) %}
+    _UniffiHelpers.RustCallWithError({{ e|error_converter_name}}.INSTANCE,
+    {%- else %}
+    _UniffiHelpers.RustCall(
+    {%- endmatch %} (ref UniffiRustCallStatus _status) =>
+    _UniFFILib.{{ func.ffi_func().name() }}(
+        thisPtr, {% call lower_arg_list(func) -%}{% if func.arguments().len() > 0 %},{% endif %} ref _status)
+)
+{%- endmacro -%}
 
 {%- macro lower_arg_list(func) %}
     {%- for arg in func.arguments() %}
@@ -54,26 +65,11 @@
     {%- endfor %}
 {%- endmacro %}
 
-{% macro arg_list_protocol(func) %}
-    {%- for arg in func.arguments() -%}
-        {{ arg|type_name(ci) }} {{ arg.name()|var_name -}}
-        {%- if !loop.last %}, {% endif -%}
-    {%- endfor %}
-{%- endmacro %}
 {#-
 // Arglist as used in the _UniFFILib function declations.
 // Note unfiltered name but ffi_type_name filters.
 -#}
 {%- macro arg_list_ffi_decl(func) %}
-    {%- if func.is_object_clone_function() %}
-    IntPtr @ptr,
-    {%- if func.has_rust_call_status_arg() %}ref UniffiRustCallStatus _uniffi_out_err{% endif %}
-    {%- else %}
-    {%- call arg_list_ffi_decl_xx(func) %}
-    {%- endif %}
-{%- endmacro -%}
-
-{%- macro arg_list_ffi_decl_xx(func) %}
     {%- for arg in func.arguments() %}
         {{- arg.type_().borrow()|arg_type_name }} {{ arg.name()|var_name -}}{%- if !loop.last || func.has_rust_call_status_arg() -%},{%- endif -%}
     {%- endfor %}
@@ -94,12 +90,6 @@
             {{ prefix }}.{{ field.name()|var_name }}{% if !loop.last %},{% endif %}
         {%- endfor %});
 {%- endmacro -%}
-
-{%- macro ffi_function_definition(func) %}
-fun {{ func.name()|fn_name }}(
-    {%- call arg_list_ffi_decl(func) %}
-){%- match func.return_type() -%}{%- when Some with (type_) %}: {{ type_|ffi_type_name }}{% when None %}: Unit{% endmatch %}
-{% endmacro %}
 
 {%- macro method_throws_annotation(throwable_type) %}
     {%- match throwable_type -%}
