+++ fixtures/custom-types-builtin/Cargo.toml
@@ -12,8 +12,8 @@ name = "uniffi_cs_custom_types_builtin"
 once_cell = "1.12"
 paste = "1.0"
 thiserror = "1.0"
-uniffi = {path = "../../3rd-party/uniffi-rs/uniffi", features=["build"]}
-uniffi_macros = {path = "../../3rd-party/uniffi-rs/uniffi_macros"}
+uniffi = { workspace = true, features = ["build"] }
+uniffi_macros.workspace = true
 
 [build-dependencies]
-uniffi = {path = "../../3rd-party/uniffi-rs/uniffi", features=["bindgen-tests"]}
+uniffi = { workspace = true, features = ["bindgen-tests"] }
