+++ fixtures/global-methods-class-name/Cargo.toml
@@ -11,8 +11,8 @@ name = "global_methods_class_name"
 [dependencies]
 once_cell = "1.12"
 thiserror = "1.0"
-uniffi = {path = "../../3rd-party/uniffi-rs/uniffi", features=["build"]}
-uniffi_macros = {path = "../../3rd-party/uniffi-rs/uniffi_macros"}
+uniffi = { workspace = true, features = ["build"] }
+uniffi_macros.workspace = true
 
 [build-dependencies]
-uniffi = {path = "../../3rd-party/uniffi-rs/uniffi", features=["bindgen-tests"]}
+uniffi = { workspace = true, features = ["bindgen-tests"] }
