+++ fixtures/optional-parameters/Cargo.toml
@@ -9,8 +9,8 @@ crate-type = ["lib", "cdylib"]
 name = "uniffi_cs_optional_parameters"
 
 [dependencies]
-uniffi = {path = "../../3rd-party/uniffi-rs/uniffi", features=["build"]}
-uniffi_macros = {path = "../../3rd-party/uniffi-rs/uniffi_macros"}
+uniffi = { workspace = true, features = ["build"] }
+uniffi_macros.workspace = true
 
 [build-dependencies]
-uniffi = {path = "../../3rd-party/uniffi-rs/uniffi", features=["bindgen-tests"]}
+uniffi = { workspace = true, features = ["bindgen-tests"] }
