+++ fixtures/regressions/issue-28/Cargo.toml
@@ -10,8 +10,8 @@ crate-type = ["lib", "cdylib"]
 name = "issue_28"
 
 [dependencies]
-uniffi = {path = "../../../3rd-party/uniffi-rs/uniffi", features=["build"]}
-uniffi_macros = {path = "../../../3rd-party/uniffi-rs/uniffi_macros"}
+uniffi = { workspace = true, features = ["build"] }
+uniffi_macros.workspace = true
 
 [build-dependencies]
-uniffi = {path = "../../../3rd-party/uniffi-rs/uniffi", features=["bindgen-tests"]}
+uniffi = { workspace = true, features = ["bindgen-tests"] }
