+++ fixtures/regressions/issue-76/Cargo.toml
@@ -9,9 +9,9 @@ crate-type = ["lib", "cdylib"]
 name = "issue_76"
 
 [dependencies]
-uniffi = {path = "../../../3rd-party/uniffi-rs/uniffi", features=["build"]}
-uniffi_macros = {path = "../../../3rd-party/uniffi-rs/uniffi_macros"}
 thiserror = "1.0"
+uniffi = { workspace = true, features = ["build"] }
+uniffi_macros.workspace = true
 
 [build-dependencies]
-uniffi = {path = "../../../3rd-party/uniffi-rs/uniffi", features=["bindgen-tests"]}
+uniffi = { workspace = true, features = ["bindgen-tests"] }
